Add hidden proxy and firmwares with the proxy compiled in.
The proxy allows direct communication with connected Bricklets.
We need direct access for the provisioning of ESP32 Bricks.
However the firmware does assume, no one except itself controls
the Bricklets, so we don't want to ship with the proxy enabled.

The hidden proxy is disabled by default, but can be enabled by
calling the API under /hidden_proxy/enable and disabled with
/hidden_proxy/disable. Calling it hidden should be seen as hint
to users who find the proxy mode: Using it is not supported and
can result in errors if the firmware's assumptions of the state
of connected Bricklets are broken.

Also add the firmware that is flashed on ESP Bricks. This does not
match the officially released version 1.2.3, because the hidden
proxy was not embedded at the time of release.