charge_manager: Only save with button.
Previously, each time a charger was added or removed,
charge_manager/config_update was called immediately and the user was
nagged to restart to apply the config.

Now only the save button (moved to the bottom to imply it will also
save the charger list) calls the API, every change done before that
only exists in the browser. This lets this configuration page
behave more like the other forms on the web interface.