evse_v2_meter: Call evse_v2.get_energy_meter_state in setup.
The EVSE V2 module is initialized earlier for the factory reset with
the wallbox button. The module setup is done so fast, that sometimes
the EVSE itself has not yet checked whether an energy meter is
available. The meter module then uses the stale information and
does not setup itself.

Updating the energy meter state when starting the meter module fixes
the race condition between the EVSE module and the bricklet itself.