evse_v2: Remove factory reset with button.
This results in a reset loop if the WARP charger is powered without
the button plugged in.

The "official" way to perform a factory reset if the web interface is
unreachable is now to remove the ESP from the charger, plug it into
a PC and do the reset with Brick Viewer.