charge_manager: Reduce switching between chargers.
We now sort the chargers twice: First by their supported current
and on whether they are already charging or not. Also we run two
passes over the chargers to calculcate the current targets. The first
run only allocates as much current to each charger as is required for
it to start or continue charging. The second pass then evenly
distributes the left-over current over all (now) active chargers.

Situations where a charger is blocked and another is unblocked should
now only happen if the available current to be distributed is changed
dynamically so that first less current is available, so a charger has
to be blocked and then later more current is available, but in the
meantime the supported current of the blocked charger was increased
so that another charger will be prioritized.