charge_manager: Only update last_sent_config of reacting chargers.
This fixes the ping-pong issue, where a charger that has an
unreactive EVSE (i.e. the sent current limits are not applied)
will then be set a target of 0 ampere, but setting this limit
increased the time the charger is given to react. It then appears
to be reactive again, as the new time limit has not immediately
elapsed.