firmwares: Shorten WARP1 index file
Due to a bug in the firmware index file handling a wrong invalid-index
error is reported if the index file is received in more than one chunk.
With the current index file being 720 byte it is received in a 708 byte
and a 12 byte chunk, triggering the problem. Removing the oldest line
reduces the size to 705 byte which should make it being received in one
chunk working around the problem.